LPS-179094 Extract ScreenNavigationEntry methods from ObjectDefinitionsDetailsScreenNavigationCategory to ObjectDefinitionsDetailsScreenNavigationEntry, implement getCategoryKey method in ObjectDefinitionsDetailsScreenNavigationEntry and make ObjectDefinitionsDetailsScreenNavigationCategory extends BaseObjectDefinitionsScreenNavigationCategory
